添加库存变更

+++ pos-item/src/main/java/com/dianba/pos/item/repository/PosItemJpaRepository.java
@@ -14,7 +14,7 @@ import java.util.List;
  */
 @Transactional
 @Repository
-public interface PosItemJpaRepository extends JpaRepository<PosItem, Integer> {
+public interface PosItemJpaRepository extends JpaRepository<PosItem, Long> {
 
 
     /**
